+++ R/ta.R
@@ -1,0 +1,3 @@
+ta <- function(x, ...)
+  amplitude(x, ...)
+