+++ demo/00-index.R
@@ -6,6 +6,7 @@
 ## source("demo/grouped-stats.R")
 ## source("demo/measures-from-fdt.R")
 ## source("demo/make-fdt.R")
-## source("demo/plot-and-xtable.R")
+## source("demo/plot.R")
+## source("demo/xtable.R")
 
 message("fdth demos created. Use source('demo/<file>.R') to run.")